pyproject.toml
[tool.poetry]
name = "dlp-lab"
version = "0.1.0"
description = ""
authors = ["ra890927 <happy.890927410@gmail.com>"]
readme = "README.md"
packages = [{include = "dlp_lab"}]

[tool.poetry.dependencies]
python = "3.9.17"
matplotlib = "^3.7.2"
torch = "^2.0.1"
torchvision = "^0.15.2"
torchaudio = "^2.0.2"
tqdm = "^4.65.0"
autopep8 = "^2.0.2"
pillow = "^10.0.0"
pandas = "^2.0.3"
opencv-python = "^4.8.0.76"
tensorboard = "^2.14.0"
gym = {version = "0.15.7", extras = ["atari"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
